Added fix for decrypting config files
We should be able to decrypt encrypted values even without knowing the
kms key, so we removed the check to see if it was present.  We also
added better ipython completion by fleshing out the `__dir__` function.
Now we provide completion both for functions as well as newly defined
values.